cidade-conectada/stolen-service/requirements.txt
google-cloud-datastore
tornado
